+++ src/toolkit/objectives.ts
@@ -280,4 +280,24 @@ function objectiveArrayToObject(objectives: Objective[]): { [key: string]: Objec
     return objectivesObject;
 }
 
-export { getObjectives, deleteObjective, markObjectiveAsDone, clearObjectives, getObjectiveByIdentifier, defineObjectiveDescription, registerBackgroundObjectivesFetchAsync, checkForTodaysObjectives, objectiveArrayToObject };
+/**
+ * Calculates the duration of each fragment of a session. Let me explain: Sessions support rests, which - as the name implies - are pauses of a fixed duration between a session, for the user to rest.
+ * While the duration of a rest is specified by the user, it's position is not, instead PersonaPlus will (thanks to this function) distribute evenly each rest between all the duration of the session. This implies splitting the session's duration into **fragments**, separated by rests.
+ *
+ * @param {number | null} duration The duration (in seconds) of the whole session.
+ * @param {number | null} rests The amount of rests of the session.
+ * @returns {number} A number, the amount of seconds each fragment shall last. If any of the params is null / invalid, throws an error.
+ */
+function calculateSessionFragmentsDuration(duration: number | null, rests: number | null): number {
+    if (rests !== null && rests < 0) {
+        termLog("React error: Negative rests? Seriously? Check your code", "error");
+        throw new Error("Negative rests? Seriously? Check your code"); // heh~
+    } else if (rests === null || duration === null) {
+        termLog("React error: Got a null value. Objective is not getting fetched correctly? Check your code", "error");
+        throw new Error("React error: Got a null value. Objective is not getting fetched correctly? Check your code"); // heh~
+    } else {
+        return duration / (rests + 1);
+    }
+}
+
+export { getObjectives, deleteObjective, markObjectiveAsDone, clearObjectives, getObjectiveByIdentifier, defineObjectiveDescription, registerBackgroundObjectivesFetchAsync, checkForTodaysObjectives, objectiveArrayToObject, calculateSessionFragmentsDuration };